refactor: extract insights breakdown card
