+++ frontend/src/insights/InsightsPage.tsx
@@ -18,8 +18,6 @@ import {
   Bar,
   BarChart,
   Cell,
-  Pie,
-  PieChart,
   ReferenceLine,
   ResponsiveContainer,
   Tooltip,
@@ -36,9 +34,15 @@ import {
 } from '@/api/insights'
 import { AppSlotMachineText } from '@/components/AppSlotMachineText'
 import { SavingsCurrentBoundary } from '@/dashboard/components/SavingsCurrentBoundary'
-import { BREAKDOWN_COLORS } from '@/dashboard/constants/breakdownColors'
 import { formatCurrency } from '@/utils/formatCurrency'
 import { useAuth } from '@/hooks/useAuth'
+import {
+  IncomeExpenseBreakdownCard,
+  type BreakdownEntry,
+  type BreakdownMode,
+  type CategoryDriver,
+  type CategoryTrendSection,
+} from './components/IncomeExpenseBreakdownCard'
 import { InsightsRangeSelector, type InsightsRangeSelectorOption } from './components/InsightsRangeSelector'
 import {
   IncomeExpenseSankeyCard,
@@ -47,14 +51,7 @@ import {
 } from './components/IncomeExpenseSankeyCard'
 import { PeriodGlanceCard } from './components/PeriodGlanceCard'
 
-type BreakdownMode = 'expense' | 'income'
 type InsightsRangePreset = 'THIS_WEEK' | 'THIS_MONTH' | 'THIS_YEAR' | 'LAST_WEEK' | 'LAST_MONTH' | 'CUSTOM'
-
-type BreakdownEntry = {
-  id: string
-  name: string
-  amount: number
-}
 
 type MerchantBubble = {
   id: string
@@ -83,14 +80,6 @@ type InsightScaffoldData = {
   incomeBreakdown: BreakdownEntry[]
   expenseBreakdown: BreakdownEntry[]
   merchantBubbles: MerchantBubble[]
-}
-
-type CategoryDriver = {
-  name: string
-  amount: number
-  previousAmount: number
-  changePct: number
-  transactionCount: number
 }
 
 type CategoryDriverGroups = {
@@ -190,26 +179,6 @@ const INSIGHTS_RANGE_OPTIONS: InsightsRangeSelectorOption<InsightsRangePreset>[]
   { value: 'LAST_MONTH', label: 'LM', description: 'Last month' },
   { value: 'CUSTOM', label: 'Custom' },
 ]
-
-const pieLegendContainerVariants = {
-  initial: { transition: { staggerChildren: 0.035 } },
-  enter: { transition: { staggerChildren: 0.045, staggerDirection: -1, delayChildren: 0.03 } },
-  exit: { transition: { staggerChildren: 0.035, staggerDirection: 1 } },
-} as const
-
-const pieLegendItemVariants = {
-  initial: { opacity: 0, y: 8, filter: 'blur(2px)' },
-  enter: { opacity: 1, y: 0, filter: 'blur(0px)' },
-  exit: { opacity: 0, y: 8, filter: 'blur(2px)' },
-} as const
-
-const pieLegendItemTransition = { duration: 0.24, ease: [0.16, 1, 0.3, 1] } as const
-
-const categoryTrendListVariants = {
-  initial: { transition: { staggerChildren: 0.03 } },
-  enter: { transition: { staggerChildren: 0.045, delayChildren: 0.03 } },
-  exit: { transition: { staggerChildren: 0.035, staggerDirection: -1 } },
-} as const
 
 const presetScaffoldRange: Record<Exclude<InsightsRangePreset, 'CUSTOM'>, SpendingRange> = {
   THIS_WEEK: 'WTD',
@@ -534,10 +503,6 @@ const insightDataByRange: Record<SpendingRange, InsightScaffoldData> = {
       { id: 'home-hardware', name: 'Neighbourhood Home Hardware', totalAmount: 104600, transactionCount: 12, averageAmount: 8717 },
     ],
   },
-}
-
-function getTotal(entries: BreakdownEntry[]) {
-  return entries.reduce((sum, entry) => sum + entry.amount, 0)
 }
 
 function getPct(amount: number, total: number) {
@@ -1041,17 +1006,6 @@ function getCategoryDriverGroups(drivers: CategoryDriver[]): CategoryDriverGroup
       .sort((a, b) => a.changePct - b.changePct)
       .slice(0, categoryDriverDirectionLimit),
   }
-}
-
-function getCategoryDriverColor(mode: BreakdownMode, changePct: number) {
-  if (changePct === 0) return 'var(--app-text-muted)'
-  if (mode === 'income') return changePct > 0 ? 'var(--app-positive)' : 'var(--app-negative)'
-  return changePct > 0 ? 'var(--app-negative)' : 'var(--app-positive)'
-}
-
-function getCategoryDriverDescriptor(changePct: number) {
-  if (changePct === 0) return 'flat'
-  return changePct > 0 ? 'increase' : 'decrease'
 }
 
 function getPeriodGlanceBrief(data: InsightsPeriodGlanceResponse, displayCurrency: string): PeriodBrief {
@@ -2038,7 +1992,6 @@ function MerchantMarketMap({
 }
 
 export default function InsightsPage() {
-  const shouldReduceMotion = useReducedMotion()
   const { user } = useAuth()
   const defaultCustomRange = useMemo(() => getDefaultCustomRange(), [])
   const [rangePreset, setRangePreset] = useState<InsightsRangePreset>('THIS_MONTH')
@@ -2062,7 +2015,6 @@ export default function InsightsPage() {
   const data = insightDataByRange[range]
   const displayCurrency = user?.base_currency ?? 'CAD'
   const selectedBreakdown = breakdownMode === 'expense' ? data.expenseBreakdown : data.incomeBreakdown
-  const selectedTotal = getTotal(selectedBreakdown)
   const flowData = useMemo(() => getFlowData(incomeExpenseFlowQuery.data), [incomeExpenseFlowQuery.data])
   const flowIncomeSources = incomeExpenseFlowQuery.data?.income_sources ?? []
   const flowExpenseCategories = incomeExpenseFlowQuery.data?.expense_categories ?? []
@@ -2116,7 +2068,7 @@ export default function InsightsPage() {
   const savingsRateHistory = useMemo(() => getSavingsRateHistory(), [])
   const merchantMarketLayout = useMemo(() => getMerchantMarketLayout(data.merchantBubbles, range), [data.merchantBubbles, range])
   const rankedMerchants = [...data.merchantBubbles].sort((a, b) => b.totalAmount - a.totalAmount)
-  const selectedCategoryTrendSections = [
+  const selectedCategoryTrendSections: CategoryTrendSection[] = [
     { id: 'increases', label: 'Top Increases', drivers: selectedCategoryDriverGroups.increases },
     { id: 'decreases', label: 'Top Decreases', drivers: selectedCategoryDriverGroups.decreases },
   ]
@@ -2167,168 +2119,35 @@ export default function InsightsPage() {
           emptyLabel={flowEmptyLabel}
         />
 
-        <section className="app-card">
-          <SectionHeader
-            icon={PieChartIcon}
-            label={(
-              <span className="inline-flex items-baseline whitespace-nowrap">
-                <AppSlotMachineText text={breakdownMode === 'expense' ? 'Expense' : 'Income'} />
-                <span className="ml-[0.25em]">Breakdown</span>
-              </span>
-            )}
-            action={(
-              <button
-                type="button"
-                onClick={() => setBreakdownMode((mode) => (mode === 'expense' ? 'income' : 'expense'))}
-                title={breakdownMode === 'expense' ? 'Show income breakdown' : 'Show expense breakdown'}
-                aria-label={breakdownMode === 'expense' ? 'Show income breakdown' : 'Show expense breakdown'}
-                className="app-icon-button"
-              >
-                <Repeat size={12} />
-              </button>
-            )}
-          />
-          <div className="grid gap-6 min-[1180px]:grid-cols-[minmax(0,0.95fr)_minmax(360px,1.05fr)]">
-            <div className="flex min-h-[620px] flex-col">
-              <div className="relative h-[450px] shrink-0">
-                <div className="pointer-events-none absolute inset-0 z-10 flex flex-col items-center justify-center">
-                  <span className="app-label app-label-compact">
-                    Total {breakdownMode === 'expense' ? 'Expense' : 'Income'}
-                  </span>
-                  <span className="font-financial text-3xl leading-none tracking-tight">
-                    {formatCurrency(selectedTotal, displayCurrency)}
-                  </span>
-                </div>
-                <ResponsiveContainer width="100%" height="100%">
-                  <PieChart>
-                    <Pie
-                      data={selectedBreakdown}
-                      cx="50%"
-                      cy="50%"
-                      innerRadius="62%"
-                      outerRadius="90%"
-                      paddingAngle={3}
-                      dataKey="amount"
-                      nameKey="name"
-                      stroke="none"
-                    >
-                      {selectedBreakdown.map((entry, index) => (
-                        <Cell key={entry.id} fill={BREAKDOWN_COLORS[index % BREAKDOWN_COLORS.length]} />
-                      ))}
-                    </Pie>
-                    <Tooltip
-                      wrapperClassName="app-chart-tooltip-default"
-                      formatter={(value, name) => [formatCurrency(Number(value), displayCurrency), name]}
-                    />
-                  </PieChart>
-                </ResponsiveContainer>
-              </div>
-              <div className="relative mt-auto min-h-[136px] overflow-hidden">
-                <AnimatePresence initial={false} mode="wait">
-                  <motion.div
-                    key={`${breakdownMode}-${range}`}
-                    className="absolute inset-x-5 bottom-0 space-y-2"
-                    variants={shouldReduceMotion ? undefined : pieLegendContainerVariants}
-                    initial={shouldReduceMotion ? false : 'initial'}
-                    animate={shouldReduceMotion ? { opacity: 1 } : 'enter'}
-                    exit={shouldReduceMotion ? undefined : 'exit'}
-                  >
-                    {selectedBreakdown.slice(0, 5).map((entry, index) => (
-                      <motion.div
-                        key={entry.id}
-                        className="flex items-center gap-3 text-sm"
-                        variants={shouldReduceMotion ? undefined : pieLegendItemVariants}
-                        transition={shouldReduceMotion ? { duration: 0 } : pieLegendItemTransition}
-                      >
-                        <span
-                          className="h-2.5 w-2.5 rounded-full"
-                          style={{ background: BREAKDOWN_COLORS[index % BREAKDOWN_COLORS.length] }}
-                        />
-                        <span className="min-w-0 flex-1 truncate" style={{ color: 'var(--app-text-muted)' }}>
-                          {entry.name}
-                        </span>
-                        <span className="font-financial">
-                          {getPct(entry.amount, selectedTotal)}%
-                        </span>
-                      </motion.div>
-                    ))}
-                  </motion.div>
-                </AnimatePresence>
-              </div>
-            </div>
-
-            <div className="flex min-h-[620px] flex-col border-t border-[var(--app-border)] pt-4 min-[1180px]:border-t-0 min-[1180px]:pt-0">
-              <div className="grid min-h-0 flex-1 grid-rows-2 gap-4">
-                {selectedCategoryTrendSections.map((section) => (
-                  <div
-                    key={section.id}
-                    className="flex min-h-0 flex-col"
-                  >
-                    <p className="app-label mb-2">
-                      {section.label}
-                    </p>
-                    <div className="min-h-0 flex-1 overflow-hidden">
-                      <AnimatePresence initial={false} mode="wait">
-                        <motion.div
-                          key={`${section.id}-${breakdownMode}-${range}`}
-                          className="space-y-2"
-                          variants={shouldReduceMotion ? undefined : categoryTrendListVariants}
-                          initial={shouldReduceMotion ? false : 'initial'}
-                          animate={shouldReduceMotion ? { opacity: 1 } : 'enter'}
-                          exit={shouldReduceMotion ? undefined : 'exit'}
-                        >
-                          {section.drivers.length === 0 ? (
-                            <motion.p
-                              key={`${section.id}-empty`}
-                              className="rounded-md border border-[var(--app-border)] px-3 py-2.5 text-sm"
-                              style={{ color: 'var(--app-text-muted)' }}
-                              variants={shouldReduceMotion ? undefined : pieLegendItemVariants}
-                              transition={shouldReduceMotion ? { duration: 0 } : pieLegendItemTransition}
-                            >
-                              No {section.id === 'increases' ? 'increases' : 'decreases'} in this period.
-                            </motion.p>
-                          ) : section.drivers.map((driver) => {
-                            const driverColor = getCategoryDriverColor(breakdownMode, driver.changePct)
-                            const changeAmount = driver.amount - driver.previousAmount
-                            return (
-                              <motion.div
-                                key={driver.name}
-                                className="grid grid-cols-[minmax(0,1fr)_auto] items-center gap-4 rounded-md border border-[var(--app-border)] px-3 py-2.5"
-                                variants={shouldReduceMotion ? undefined : pieLegendItemVariants}
-                                transition={shouldReduceMotion ? { duration: 0 } : pieLegendItemTransition}
-                              >
-                                <div className="min-w-0">
-                                  <p className="truncate font-semibold">
-                                    {driver.name}
-                                  </p>
-                                  <p className="text-sm" style={{ color: 'var(--app-text-muted)' }}>
-                                    {driver.transactionCount} transactions | previous {formatCurrency(driver.previousAmount, displayCurrency)}
-                                  </p>
-                                </div>
-                                <div className="text-right">
-                                  <p className="font-financial text-base">
-                                    {formatCurrency(driver.amount, displayCurrency)}
-                                  </p>
-                                  <p className="mt-1 font-financial text-sm" style={{ color: driverColor }}>
-                                    {formatSignedCurrency(changeAmount, displayCurrency)}
-                                    {' '}
-                                    ({driver.changePct > 0 ? '+' : ''}{driver.changePct}%)
-                                    {' '}
-                                    {getCategoryDriverDescriptor(driver.changePct)}
-                                  </p>
-                                </div>
-                              </motion.div>
-                            )
-                          })}
-                        </motion.div>
-                      </AnimatePresence>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </section>
+        <IncomeExpenseBreakdownCard
+          header={(
+            <SectionHeader
+              icon={PieChartIcon}
+              label={(
+                <span className="inline-flex items-baseline whitespace-nowrap">
+                  <AppSlotMachineText text={breakdownMode === 'expense' ? 'Expense' : 'Income'} />
+                  <span className="ml-[0.25em]">Breakdown</span>
+                </span>
+              )}
+              action={(
+                <button
+                  type="button"
+                  onClick={() => setBreakdownMode((mode) => (mode === 'expense' ? 'income' : 'expense'))}
+                  title={breakdownMode === 'expense' ? 'Show income breakdown' : 'Show expense breakdown'}
+                  aria-label={breakdownMode === 'expense' ? 'Show income breakdown' : 'Show expense breakdown'}
+                  className="app-icon-button"
+                >
+                  <Repeat size={12} />
+                </button>
+              )}
+            />
+          )}
+          mode={breakdownMode}
+          entries={selectedBreakdown}
+          trendSections={selectedCategoryTrendSections}
+          displayCurrency={displayCurrency}
+          animationKey={`${breakdownMode}-${range}`}
+        />
 
         <section className="app-card">
           <SectionHeader icon={Wallet} label="Net Worth" />
